[Store] Move BatchEvict benchmark to benchmarks/ as standalone executable
Per review feedback, move the BatchEvict benchmark out of the unit-test
suite into mooncake-store/benchmarks/, wired through the benchmark CMake
path. Rewritten from a gtest fixture into a standalone main() program
following the existing benchmark layout (glog/gflags, explicit checks).
Friend declaration updated from test::BatchEvictBenchTest to
benchmarks::BatchEvictBench.
