+++ mooncake-store/include/master_service.h
@@ -58,8 +58,10 @@ class SnapshotChildProcessTest;
 // exposing test-only accessors on MasterService itself.
 class PromotionOnHitTest;
 class MasterServiceTenantQuotaTest;
-class BatchEvictBenchTest;
 }  // namespace test
+namespace benchmarks {
+class BatchEvictBench;
+}  // namespace benchmarks
 
 /*
  * @brief MasterService is the main class for the master server.
@@ -74,7 +76,7 @@ class MasterService {
     friend class test::MasterServiceSnapshotTestBase;
     friend class test::SnapshotChildProcessTest;
     friend class test::PromotionOnHitTest;
-    friend class test::BatchEvictBenchTest;
+    friend class benchmarks::BatchEvictBench;
     friend class test::MasterServiceTenantQuotaTest;
 
    public:
